+++ bot/cogs/admin.py
@@ -1,12 +1,15 @@
 from discord.ext import commands
 
 import modules.config as cfg
-from modules.enumerations import SelStatus, MatchStatus
-from modules.display import send
+from modules.enumerations import SelStatus, MatchStatus, PlayerStatus
+from modules.display import send, channelSend
 from modules.exceptions import ElementNotFound
 from modules.tools import isAdmin
+from modules.database import remove
 
-from matches import clearLobby, getMatch, getAllNamesInLobby
+from classes.players import removePlayer, getPlayer
+
+from matches import clearLobby, getMatch, getAllNamesInLobby, removeFromLobby
 
 
 
@@ -76,6 +79,34 @@ class AdminCog(commands.Cog, name='admin'):
             match.confirmMap()
             await send("MATCH_MAP_SELECTED", ctx, sel.map.name)
 
+    @commands.command()
+    @commands.guild_only()
+    async def unregister(self, ctx):
+        if ctx.channel.id != cfg.discord_ids["register"]:
+            await send("WRONG_CHANNEL", ctx, ctx.command.name, f"<#{cfg.discord_ids['register']}>")
+            return
+        if len(ctx.message.mentions) != 1:
+            await send("RM_MENTION_ONE", ctx)
+            return
+        try:
+            player = getPlayer(ctx.message.mentions[0].id)
+        except ElementNotFound:
+            await send("RM_NOT_IN_DB", ctx) # player isn't even registered in the system...
+            return
+        if player.status == PlayerStatus.IS_LOBBIED:
+            removeFromLobby(player)
+            await channelSend("RM_LOBBY", cfg.discord_ids["lobby"], player.mention, namesInLobby=getAllNamesInLobby())
+        if player.status == PlayerStatus.IS_REGISTERED:
+            await remove(player)
+            memb = ctx.author.guild.get_member(player.id)
+            removePlayer(player.id)
+            notify = memb.guild.get_role(cfg.discord_ids["notify_role"])
+            registered = memb.guild.get_role(cfg.discord_ids["registered_role"])
+            await memb.remove_roles(notify)
+            await memb.remove_roles(registered)
+            await send("RM_OK", ctx)
+            return
+        await send("RM_IN_MATCH", ctx)
 
 def setup(client):
     client.add_cog(AdminCog(client))