+++ bot/main.py
@@ -215,5 +215,5 @@ def main(launchStr=""):
 if __name__ == "__main__":
     # execute only if run as a script
     # Use main() for production
-    # main("_test")
-    main()
+    main("_test")
+    #main()